Make linux-win_cross-rel a tot bot for real.
I wanted to do this so we check that the cross bot keeps working at tot.
But it's also needed to disable goma for the bot, which doesn't work in
cross builds with pinned clang either (the bot currently tries to use
goma and hence has a very slow compile).

Bug: 495204
Change-Id: I5da4fea60902508ffc5d5b76919d6772cf46bf6e
Reviewed-on: https://chromium-review.googlesource.com/933341
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>