V8: Extend v8 deps roller to roll test262
This makes the v8-deps-roller recipe more generic and allows running
different versions of it on different builders.

The previous configuration is refactored, but remains the same. A new
configuration for solely rolling test262 is added. The new roller will
run on a separate builder, sharing the same slave.

Bug: v8:7419
Change-Id: Ic88e65b233f6b545186b32e9a115cb4621329855
Reviewed-on: https://chromium-review.googlesource.com/939855
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>