Temporarily disable isac_fix_test in the bot config.
isac_fix_test used to break with this bot config (which makes
webrtc_perf_tests work), but was fixed in the WebRTC code. However,
in order to zero in on a regression, we'd like to be able to run old
jobs with the new bot config. We can't do that because of
isac_fix_test, so let's disable it for a day or two while we bisect.

TBR=ehmaldonado@chromium.org
Bug: chromium:817310
Change-Id: Ibf5274a7770ee469d2704692fa1df5ce85a246c8
Reviewed-on: https://chromium-review.googlesource.com/941943
Commit-Queue: Patrik Höglund <phoglund@chromium.org>
Reviewed-by: Patrik Höglund <phoglund@chromium.org>