Use canary version of goma in chromeos goma-canary builder
I've set up amd64-generic-goma-canary-chromium-pfq-informational builder
in master chromiumos.chromium, and it looks working.
https://ci.chromium.org/buildbot/chromiumos.chromium/amd64-generic-goma-canary-chromium-pfq-informational/

The last step is to use canary version of goma in that builder.

I was requested to use property before, but buildbot property is
basically set by buildbot. In chrome, we have a map from builder name to
config, so I think it's ok to follow this way in chromeos side.

Bug: chromium:746816
Change-Id: I57a7b8b23fc9b53d1f3729fc00db0537250abcc5
Reviewed-on: https://chromium-review.googlesource.com/939143
Commit-Queue: Shinya Kawanaka <shinyak@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>