Simplify perf test uploading + squelch NOT_UPLOADING_CHART warnings.
I believe this makes the code a bit less hacky and clears up
what's really going on.

This moves step_metadata to the upload step rather than the test
step. I think this is strange, but Chromium is doing it like that:

https://ci.chromium.org/buildbot/chromium.perf/Android%20Nexus5X%20Perf/1350

AFAIK we're note using step metadata anyway; it appears to be used
for findit, which we're not using.

There are quite a few follow up things to investigate here, like
1) can we move off our custom upload mechanism and use chromium's,
2) why do our regular tests not use SwarmingGTestTest?
3) why do NOT_UPLOADING_CHART_JSON not show up in the recipe json
   files, while they do on the bots? I think this is because the
   swarming collect step isn't stubbed to return the right thing,
   so the test believes there are no results and we don't go into
   the not-uploading branch. I spent a lot of time trying to figure
   out how to stub this, but to no avail.

Bug: webrtc:8961
Change-Id: Ib868e9464b389bb1901ea3c64663c22e9bbdb906
Reviewed-on: https://chromium-review.googlesource.com/943108
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Commit-Queue: Patrik Höglund <phoglund@chromium.org>