Fix ios_framework_api bot.
This recipe does a lot of things the ios recipe does, but it doesn't
build using the ios recipe code. This fixes the code so that the
correct xcode is downloaded. This removes FORCE_MAC_TOOLCHAIN
since the cipd mechanism supersedes it.

Possible follow-up: put a .json next to the others and make this
recipe read it, so the xcode version is just in one place.

Bug: chromium:821309
Change-Id: Id9766febfc054698f8d4e2cc6eac6433df67cd8b
Reviewed-on: https://chromium-review.googlesource.com/960030
Reviewed-by: Sergiy Byelozyorov <sergiyb@chromium.org>
Commit-Queue: Patrik Höglund <phoglund@chromium.org>