[chromium_checkout] Change how the checkout dir is calculated on LUCI.
Previously the checkout_dir would be calculated like:

    buildbot: \b\c\b\GPU_FYI_Win_Clang_Builder__dbg_
    buildbot: \b\c\b\<checkout_dir>
    luci:     \b\swarming\w\ir\cache\builder\GPU_FYI_Win_Clang_Builder__dbg_
    luci:     \b\swarming\w\ir\cache\builder\<checkout_dir>

On Buildbot 'api.path["cache"]' was shared between all builders that shared
the bot. On LUCI the 'cache\builder' subfolder of the task is uniquely
generated per builder. Thus the buildername, which was previously a required
disambiguation component, is now actively harmful.

To further complicate things, the `bot_config` in the recipes allows the
specification of a "checkout_dir". This feature was intended to allow
multiple builders to share the same checkout directory. However, because
'cache\builder' is unique per builder in LUCI, no such sharing occurs.

This CL changes the path calculation to:

    buildbot: \b\c\b\GPU_FYI_Win_Clang_Builder__dbg_
    buildbot: \b\c\b\<checkout_dir>
    luci:     \b\swarming\w\ir\cache\builder

The "builder" cache can be overridden in a cr-buildbucket cache entry
from its default in the event that you want multiple bots to share the
same named cache.

R=jbudorick@chromium.org, nodir@chromium.org, tandrii@chromium.org

Bug: 820532
Change-Id: Id14f75cd35e420f8529c7a9a105426a29b3301a9
Reviewed-on: https://chromium-review.googlesource.com/959862
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>