gatekeeper: Handle milo giving inconsistent data
It appears milo sometimes gives us different (roughly) empty values for
urls for a step in a build. Sometimes it gives a "{}" and sometimes a
"null". Dart builds somehow return null instead of {}, which means that
the chromium_build app sees this (JS) null, which is converted to a
python None, and gets confused.

Additionally, change the default value to be a dict, since the mailer
https://crrev.com/4abc8b9d659ea87b757239f418b0ced08f866b50/appengine/chromium_build/templates/waterfall_mail.html#36
seems to expect a dict.

Bug: 822050
Change-Id: I6625d3b1f78ec1defc37de54112d525c2633bddd
Reviewed-on: https://chromium-review.googlesource.com/963751
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>