Rename Mac Views builder and consolidate optional trybot pools.
The Views builder configs had "10.10" in the name, but weren't running
on 10.10. This CL renames them to "mac-views-rel", following the new
bit.ly/chromium-build-naming standard.

It also consolidates a bunch of optional builders into a single pool.
We used to have multiple pools because they were running different OS
version and because we had concerns over too many checkout per machine,
but now all of the builders use the same checkout name and are running
the same version.

We should get rid of the *variable* builders as well, and consolidate
those builders into the same pool, but that should be a follow-up CL.

R=jbudorick@chromium.org, lindsayw@chromium.org, spqchan@chromium.org
BUG=815222

Change-Id: If4701062176fa42b9f901811054a4749de52663f
Reviewed-on: https://chromium-review.googlesource.com/969794
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>