Promote GCC bot to main wf + cq.
For this bot to have any effect, it needs to block CLs in //build that
breaks GCC, so let's move it in there. Alternatively, we could keep
the bot in .fyi and just add it to the CQ, not sure what the
convention is here.

Bug: chromium:776284
Change-Id: Iff0a7a05ccf3ddcb9424b4785dc64be10d1c8b5d
Reviewed-on: https://chromium-review.googlesource.com/958902
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>