Moving packaging builds to use master for script source.
We've been building packages in the same branch as the desired output,
but this is a problem, since if the packaging script changes its output,
it can't read or write the 'latest' metadata format from gcloud
until those changes make it into the branch, which could be a month or
more for beta.

This CL makes the packaging builds sync the master branch to run the
packaging script from, and stops running tests on the packaging builds
altogether, since those refs have already been tested when they went
through master.

Bug: None.
Change-Id: I8fe471ae85f86becb747299f315f1e664b66ba68
Reviewed-on: https://chromium-review.googlesource.com/989404
Reviewed-by: Michael Goderbauer <goderbauer@google.com>
Commit-Queue: Greg Spencer <gspencer@google.com>