Fix codesearch initiator to strip prefix from json
Gerrit JSON responses all start with a well-known prefix to combat
XSS exploits. The get_json() call needs to strip that prefix
before trying to parse the response. Luckily, this is a thing
we've run into many times before, so we already have support for
it built in to the url api!

R=flyboy, jsca

Bug: 827022
Change-Id: Iddb462cb815177d7710dbbd7d0618d51c1e3a7f5
Reviewed-on: https://chromium-review.googlesource.com/996400
Reviewed-by: Alan Bram <flyboy@chromium.org>
Commit-Queue: Aaron Gable <agable@chromium.org>