Run a bit less content_unittests on WebRTC bots.
RenderWidgetHostViewAuraTest.ConflictingAllocationsResolve is
currently breaking our WebRTC Win7 bot, and that isn't helpful for us.
It happens to work on all current Chromium bots, but not ours.

We have historically just run all content_unittests because they're
fast, and it's a bit tricky to construct an appropriate filter. Now
content_unittests has grown to ~60 seconds execution time and this
unrelated test breaks our bot, so let's apply some filtering. I
constructed the filter from:

https://cs.chromium.org/chromium/src/content/test/BUILD.gn?l=1819

The filter doesn't have to be perfect as long as it covers most of
the interesting cases. If we miss something the roll will fail, and
then it will become apparent we need to update the filter. I reckon
this is extremely unlikely though.

Bug: None
Change-Id: Ie1b8dfc17e049f332e7e73f20e8613e6a4da0c29
Reviewed-on: https://chromium-review.googlesource.com/1011607
Reviewed-by: Oleh Prypin <oprypin@chromium.org>
Commit-Queue: Patrik Höglund <phoglund@chromium.org>