master.chromiumos.chromium: Remove all builds.
This change removes all of the builds currently run on the
chromiumos.chromium waterfall, but does not delete the all of the
configuration information.

They have already been duplicated on LUCI swarming, which means the
builds here are redundant. This will be followed shortly by a request
to shutdown and remove the waterfall entirely.

BUG=chromium:829169

Change-Id: I2c73fd9c9136995c4e3588a675c3db77559757fc
Reviewed-on: https://chromium-review.googlesource.com/1012862
Reviewed-by: Ningning Xia <nxia@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>