codesearch: don't upload xref index pack if compile fails
When the codesearch gen builders fail the compile step, the index
pack they produce contains fewer crossreferences than normal. This
is expected, because the compilation units which failed to compile
haven't produced any indexable objects.

Historically, when compile fails, we upload the index pack anyway.
This was because it was considered better for Grok to have an
incomplete set of xrefs than a stale set of xrefs.

However, Kythe has a validation stage which makes sure that the
xref index hasn't massively changed. When we upload an index pack
from a failed compile, Kythe's validation pipeline fails, and the
new index doesn't get served anyway. So by uploading incomplete
indices, we're not getting recent-but-partial results served,
we're just creating noise in the pipeline and continuing to serve
stale results.

This CL makes it so that we abort and don't upload an index pack
when compile fails. This will have the same end result (stale
xrefs due to failed compile), but will stop producing noisy
validation failures in the index pipeline.

This CL also renames the two other failure-mode test expectation
files to more accurately reflect the name of the failed step.

Bug: b/78093954
Change-Id: Ifb0e49b979e83dd43beca651b981abfa64e64a67
Reviewed-on: https://chromium-review.googlesource.com/1014184
Reviewed-by: Joey Scarr <jsca@chromium.org>
Commit-Queue: Aaron Gable <agable@chromium.org>