[V8] Simplify test-variant configurations
This consistently passes the --variants flag to the v8 test runner to
simplify extra corner-case logic for the default.

Previously, we didn't pass the flag on some builders, which is
equivalent to passing --variants=default.

The numfuzzer doesn't support other variants, but we add the flag here
to stay compatible: https://crrev.com/c/1015004

Bug: 830557
Change-Id: Ia4d667726e452eafcfe228803b6c839f47f7e5a9
Reviewed-on: https://chromium-review.googlesource.com/1015000
Reviewed-by: Sergiy Byelozyorov <sergiyb@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>