Don't clear cros-chrome-sdk cache on the daisy simplechrome bot.
Runhooks is taking ~15min on chromeos-daisy-rel every build because of
broken LKGM + fallback logic.

If we don't clear the cache, we skip that fallback logic, so this
should save us around 10min on every build.

Normally wouldn't be a big problem, but that slave pool is already
over-subscribed, so hopefully this is immediate relief.

Recipe-Nontrivial-Roll: build_limited_scripts_slave
Bug: 832406
Change-Id: I5f9694b184b4864c672fe24bea8acc5ae4918634
Reviewed-on: https://chromium-review.googlesource.com/1015535
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>