[V8] Refactoring - extract test spec into class
The code for processing test specs is extracted into a class and moved
into a separate file. The logic in the code is left unchanged.

This will make usage of test specs clearer on the API side, since
OO and descriptive instance methods obviate the need to document the
structures everywhere.

Bug: 830557
Change-Id: I5da557053f3b581d98633969ef8980250d23583b
Reviewed-on: https://chromium-review.googlesource.com/1016401
Reviewed-by: Sergiy Byelozyorov <sergiyb@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>