Upload images to Gold even when some tests fail for PDFium
PDFium as part of its tests has a number of suites that are performing
pixel test over a body of generated PDF docs. Currently only the 'corpus
test' results are being sent to Gold, if and only if the tests
succeeds. This is not very useful for debugging why tests are failing.

This CL changes the behaviour so that all of the related suite results
will be in the upload to Gold, and that all of the suites and upload
will run even if there are failures.

BUG=pdfium:1073

Change-Id: Ic275cd2a560f3b56c7ecd233e5458f75a65fcd09
Reviewed-on: https://chromium-review.googlesource.com/1030955
Reviewed-by: Stephan Altmueller <stephana@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>