[V8] Remove necessity to add _x64 in 64 bit windows builds
This was needed for ninja with gyp, which is obsolete now that all
builders are on gn. Removing this suffix will reduce confusion and
technical dept. We also don't need to specify the bitness anymore on
testers, which was only necessary to infer the folder name.

This was tested with led for luci trybots:
https://ci.chromium.org/swarming/task/3d3e2019d6e98610
https://ci.chromium.org/swarming/task/3d3e2a29082b5d10

Bug: 830557
Change-Id: Ib3edf6ba34b6ff2c46e810c5f2ddd9b94485c109
Reviewed-on: https://chromium-review.googlesource.com/1041845
Reviewed-by: Sergiy Byelozyorov <sergiyb@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>