[V8] Remove static build_config where it's passed as property
This also simulates the property's values based on the statically
configured parent builders.

This also removes the TARGET_BITS on testers, which is not read anymore.

In a follow up we'll soon delete most of these test cases and a generic
simulation won't be necessary any longer.

Bug: 830557
Change-Id: I82dfb4ddaf71b34495b928716e84a8633aac7b3c
Reviewed-on: https://chromium-review.googlesource.com/1042288
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Sergiy Byelozyorov <sergiyb@chromium.org>