[V8] Enable default bot_configs
This enables using the V8 recipe without a configuration for the
builder in builders.py. By default it uses an empty configuration and
only makes use of the properties passed to the recipe.

This is tested by removing one of the empty triggered-tester configs.
Others will be removed in follow ups.

Bug: 830557
Change-Id: I4260f008574bc8933a9af214180dc616d7d921a4
Reviewed-on: https://chromium-review.googlesource.com/1047425
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Sergiy Byelozyorov <sergiyb@chromium.org>