Up timeout values for linux-annotator-rel.
linux-annotator-rel runs a local script test that takes a long
time to run and doesn't produce any output. By default, on buildbot,
we'll kill a step that runs for 2400s without output, but will let
the build run indefinitely otherwise.

This CL adjusts the build to be able to run for at most three hours
total (including the build) with or without output.

Change-Id: I4648309f1821034db7147e70b79332c3f6759de3
Reviewed-on: https://chromium-review.googlesource.com/1053391
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>