Have Ninja output directly to a file instead of using raw_io.
To achieve this, I wrote a small wrapper script in Python that allows
redirecting stdout of an arbitrary command to a file.

I expect this to solve some MemoryErrors we're seeing with 32-bit
Python on Windows. Prior to this CL, the Windows builder fails when
trying to re-encode the 211MB compilation database in-memory as UTF-8.

This also means we can remove the builder step which copies
compile_commands.json into the right place, since we now write it to the right
place in the first place.

Example failure:
https://ci.chromium.org/buildbot/chromium.infra.codesearch/codesearch-gen-chromium-win/54

Bug: 840606
Change-Id: Ia51e2a4dbb3069199a172d1c29672ac0b042bca1
Recipe-Nontrivial-Roll: build_limited_scripts_slave
Recipe-Manual-Change: build_limited_scripts_slave
Reviewed-on: https://chromium-review.googlesource.com/1053039
Commit-Queue: Joey Scarr <jsca@chromium.org>
Reviewed-by: Sergey Berezin <sergeyberezin@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>