Drop swarming task priority of LUCI experimental tests.
If we are capacity constrained, currently we'll run tests
for FYI builders and LUCI migration builds at the same priority.

This CL changes the priority for the migration (experimental)
builds to be lower priority, so that they'll be starved first,
before any FYI builders.

Bug: 842940
Change-Id: I7a4845350061daef4381806d2aeaeb2b4ba855d4
Reviewed-on: https://chromium-review.googlesource.com/1064551
Reviewed-by: Ryan Tseng <hinoka@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>