Add support for custom out directories to package_index.py.
This missing feature should be the final piece of the puzzle to get
Android xrefs working in Code Search. Before this CL, there is a bug
where output is written to src/out/chromium-android/Debug, but the code
in package_index.py assumes that it's in src/out/Debug. This results
in incorrect relative paths in the Kythe index packs.

Bug: 643026
Change-Id: I39d3da00cbec6769337e43fb12856c0ad2ae8104
Reviewed-on: https://chromium-review.googlesource.com/1065962
Reviewed-by: Aaron Gable <agable@chromium.org>
Commit-Queue: Joey Scarr <jsca@chromium.org>