WebRTC: Get size of .so files without padding
This uses `readelf` to get the sum of sizes of sections in the file, rather than getting the actual size of the file. That is needed because .so files have padding and so their size is always a multiple of 4 KiB. So this will be more accurate for showing the delta in size. But note that the sum of sections will always be up to 10 KiB smaller than the size of the file.

This implementation was suggested by kwiberg@ and nisse@.

Bug: webrtc:9205
Change-Id: I79cca2e868d421c9e11586b101e0b989623f2041
Reviewed-on: https://chromium-review.googlesource.com/1071587
Reviewed-by: Patrik Höglund <phoglund@chromium.org>
Commit-Queue: Oleh Prypin <oprypin@chromium.org>