[V8] Dynamically read test roots
This makes the v8 recipe dynamically check for additional test root
directories. Usage of the test roots comes in a follow up.

This prepares for https://crrev.com/c/1086802 as it leads to a lot of
expectation changes.

This also depends on https://crrev.com/c/1088694 on V8 side.

Bug: 846711
Change-Id: I28b063ff324f6945c0d69b37fbb3f1c94f3ed3a7
Reviewed-on: https://chromium-review.googlesource.com/1093316
Reviewed-by: Sergiy Byelozyorov <sergiyb@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>