Receive the result of generate_isolated_script at last
This patch is to move webkit_layout_step to near the last of build step.
When webkit_layout_tests runs, 'webkit_layout_tests (with patch)' step is blocked to wait finish of test execution on swarming.
After that, builder receives the result of other gtest tests and uploads test result.

But, blocked time waiting webkit_layout_tests is not utilized in such build.
Instead, it is better to receive and upload the result of other tests before waiting result of webkit_layout_tests.

For example, below builds finished 'Upload to test-results [webkit_layout_tests on Intel GPU on Mac (with patch)]' in 20:55 (JST), but it finished all build steps in 21:03 (JST).
https://ci.chromium.org/p/chromium/builders/luci.chromium.try/mac_chromium_rel_ng/69248
If we move webkit_layout_tests to near the last of build, some amount of time used to wait the result of webkit_layout_tests will be used to receive and upload the results of other unittests.

Some other example builds are here.
https://ci.chromium.org/p/chromium/builders/luci.chromium.try/win7_chromium_rel_ng/12861
- webkit_layout_tests result upload step finished in 21:00 (JST), build finished in 21:12 (JST)

https://ci.chromium.org/p/chromium/builders/luci.chromium.try/win7_chromium_rel_ng/12858
- webkit_layout_tests result upload step finished in 20:33 (JST), build finished in 20:46 (JST)

https://ci.chromium.org/p/chromium/builders/luci.chromium.try/mac_chromium_rel_ng/69247
- webkit_layout_tests result upload step finished in 20:56 (JST), build finished in 21:03 (JST)

https://ci.chromium.org/p/chromium/builders/luci.chromium.try/mac_chromium_rel_ng/69245
- webkit_layout_tests result upload step finished in 21:07 (JST), build finished in 21:14 (JST)


So I expect this re-ordering removes roughly
* 12~13 mins from win7_chromium_rel_ng
* 6~8 mins from mac_chromium_rel_ng
for the builds running webkit_layout_tests with many other tests.


Bug: 851355
Change-Id: Ie5836080708f04fc16292e79fea458ba7661e980
Reviewed-on: https://chromium-review.googlesource.com/1096723
Commit-Queue: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>