Assign slave{244..247}-c1 to "Android CFI Builder Perf FYI".
Each of these machines has a different configuration:
244: n1-standard-8
245: n1-highmem-8
246: n1-standard-16
247: n1-highmem-16

This is part of an experiment to discover the minimum capacity
required to run Android ThinLTO perf bots. Once we have found the
minimum, this change will be reverted and we will update the
configuration of slave270-c1.

Bug: 821860, 853050
Change-Id: I99ad99c396a0cad6c731a7972dc22bef5d0b80aa
Reviewed-on: https://chromium-review.googlesource.com/1102178
Commit-Queue: Peter Collingbourne <pcc@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>