Remove obsolete chromium.webkit, tryserver.blink builders.
This removes a number of builders that are no longer needed given
the coverage we have on the chromium.{linux,mac,win} waterfalls
(or are just generally not needed otherwise, like the *compile*
trybots).

- chromium.webkit:
  - WebKit Linux Trusty
  - Webkit Linux Trusty (dbg)
  - WebKit Mac Builder (dbg)
  - WebKit Mac 10.10
  - WebKit Mac 10.11
  - WebKit Mac 10.11 (dbg)
  - WebKit Mac 10.12
  - WebKit Win Builder (dbg)
  - WebKit Win7
  - WebKit Win7 (dbg)
  - WebKit Win x64 Builder
  - WebKit Win x64 Builder (dbg)
- tryserver.blink:
  - linux_trusty_blink_compile_dbg
  - linux_trusty_blink_compile_rel
  - linux_trusty_blink_dbg
  - win7_blink_compile_dbg
  - win7_blink_compile_rel
  - win7_blink_dbg

This removes all of the debug builders, as they are covered on
the main waterfalls, and removes all of the redundant release
builders. There are still two builders that remain: the retina
builder and the Win10 32-bit builder.

R=hinoka@chromium.org, robertma@chromium.org

Change-Id: Ib4823772f0525bec2f6725ab55df8d05c074b058
Reviewed-on: https://chromium-review.googlesource.com/1096378
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Ryan Tseng <hinoka@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>