Add tests to ensure that the builder and trybot specs are valid.
A test case is created for each builder in the BUILDERS variable that
creates a BotConfig and makes sure that the builder can be configured.

A test case is created for each trybot in the TRYBOTS variable that sets
the mastername and buildername properties to the appropiate values and
makes sure that the trybot steps can be run.

The builder 'Android Cronet Builder Asan' specifies an invalid
combination of chromium configs. Attempting to fix it was resulting in
rippling changes, so instead the test blacklists that builder so that it
is not checked currently.

Change-Id: Iea2fe571c08507f4a9b7c0401486983a48046295
Reviewed-on: https://chromium-review.googlesource.com/1101709
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>