isolate: update wrapper script to use new isolate Go tool under src/
With the move to distribution via CIPD, it's now always downloaded at the same
place, and not inside a directory named against the OS.

Keep the old location working too, so that old checkouts are not (yet) broken.

Do not fix the iOS recipe yet, will do in a separate change.

Bug: 851596
Change-Id: Ic10959ce0f9bae8d4c452cca00d11d6a94f71c65
Reviewed-on: https://chromium-review.googlesource.com/1106006
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>