Add a trybot for Linux with DCHECKs disabled.
DCHECKs are defined as do-nothing operations when the product is
released, but all of the existing trybots enable DCHECKs so that
violations of expectations can be surfaced early. However, if the
expression used as a condition has important side effects, then
disabling DCHECKs changes the meaning of the code and this won't be
discovered until the product is actually released.

Bug: 833482
Change-Id: Ia71fa8e190caeb362ec2c8513ccdad683a2dcc17
Reviewed-on: https://chromium-review.googlesource.com/1114228
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>