Add support to CS recipe for custom out directories.
This CL enables compilations for the Code Search builders to be run in a
directory that looks like:
src/out/win-Debug/

Before this change, the deepest subdir had to be called "Debug" or some
other build config name.

This change will allow us to create generated files with correct
cross-references for different platforms in different directories,
without changing the depth of the output dir (on which the Windows
builder depends).

Note that this change should be a no-op. I'll follow this up with a separate CL
to move the output dirs of each platform to follow this convention.


Bug: 665761
Change-Id: Id0107859f0011552e7b4da24f35d6efe9b191ede
Reviewed-on: https://chromium-review.googlesource.com/1116216
Commit-Queue: Joey Scarr <jsca@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>