[V8] Re-add support for fully-qualified ref format for branch property
It was originally added in https://crrev.com/c/1122121 and then removed again in
https://crrev.com/c/1122874 due to a wrong assumption that it was only used on
Auto-tag builder. It is actually used on all builders on the branches console,
but the tests did not reflect this correctly, which caused missing coverage.
I've modified tests to simulate LUCI environment for beta branch builders.

R=machenbach@chromium.org

Bug: 793687
Change-Id: I5813b091d30c8c96ee5d72d1ffebc42e8884f12c
Reviewed-on: https://chromium-review.googlesource.com/1124160
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Sergiy Byelozyorov <sergiyb@chromium.org>