Use the generate_compdb.py script.
We do this instead of calling ninja directly, as this scripts handles
some details like setting the driver mode when the entries use clang-cl.

Note that this no longer passes the goma recommended -j arg to ninja.
This should be fine, as generating the compilation database finishes
quickly even running locally with one thread. We should check to see if
it gets too slow though, and revert if so.

Bug: 396230
Change-Id: Ia51ac38db9d4444f241c9d9787609eeee83431d8
Recipe-Manual-Change: build_limited_scripts_slave
Reviewed-on: https://chromium-review.googlesource.com/1125558
Commit-Queue: Owen Rodley <orodley@chromium.org>
Reviewed-by: Joey Scarr <jsca@chromium.org>