Check if the test data is enabled before attempting to access it.
chromium_tests uses some module test data to control the behavior when
a bad bot ID is provided so that builders.py and trybots.py can be
sanity tested. However when actually running a recipe that has a bad
bot ID, the test data is disabled and the get call results in a
difficult to understand exception rather than producing a failing step
with helpful information.

Change-Id: Ia2a80d1d24dfb8ef283945b279d68424d4e80c6c
Reviewed-on: https://chromium-review.googlesource.com/1132331
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>