Enable remote_run and use builder cache for WASM builders
On Buildbot without remote_run, builds were always happening in the same
directory, whereas on LUCI, the directory is re-created between the runs,
thereby increasing build times. To fix this, we need to tell the recipe to use
'builder_cache' path for the checkout, however, this path is only available in
remote_run configuration on Buildbot, therefore we need to enable it to prevent
changes to the recipe break current Buildbot builders.

TBR=macchenbach@chromium.org

Bug: 861811
Change-Id: I18bfd21196a3e09d88d01b5da0eb255f112416e9
Reviewed-on: https://chromium-review.googlesource.com/1134991
Reviewed-by: Sergiy Byelozyorov <sergiyb@chromium.org>
Commit-Queue: Sergiy Byelozyorov <sergiyb@chromium.org>