Reland Remove Closure Compilation step from Closure Bot.
This CL removes the Closure Compilation step from the Closure Compilation Bot
as the third_party/closure_compiler/compiled_resources2.gyp file is now
empty, and all Closure Compilation is done on the vanilla CQ.

This Bot cannot be fully removed until the Closure Tests under run_tests on this
bot are migrated to other bots.

Relanded with the compile step removed as well, as this was breaking the bots
due to lack of a GYP generation. Original CL at
https://chromium-review.googlesource.com/c/chromium/tools/build/+/1118059.

Bug: 632206
Change-Id: I050141060e590f83bfef6992b5ae9c10acaf14f4
Reviewed-on: https://chromium-review.googlesource.com/1124206
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: calamity <calamity@chromium.org>