Fix linux-goma-rbe-staging-rel bot
Changes it to compile only, and makes it apply the correct config goma
config.

It does this by mirroring the same bot on chromium.linux, which is a
"fake" CI bot with no actual bots running builds. It's there purely for
this trybot to mirror.

Bug: 855319
Change-Id: Iaea4a590eadada1bccb564c9fa070200fab81750
Reviewed-on: https://chromium-review.googlesource.com/1142452
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>