Exclude thinlto-cache directory from package archives on Android.
This directory contains a cache of object files created by the linker
during ThinLTO. It does not need to be included in the archive because
it is used only by the linker and can reach many GBs in size.

The inclusion of the cache directory had also caused the cycle times
on the Android perf builders to increase such that they had fallen
significantly behind trunk. This change should allow them to catch up.

NOTRY=true

Bug: 469376
Change-Id: I8feac86d0938b55e64baa85f4672a07e4dc899de
Reviewed-on: https://chromium-review.googlesource.com/1161203
Commit-Queue: Peter Collingbourne <pcc@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>