[V8] Fix flake bisection progress and isolate caching
This fixes two bugs when bisecting over a large gap of missing builds.

Given a range 4..0 of commits where 4 is good and 0 is bad. There are
no builds for 1, 2, and 3.

Previously, the isolate lookup for 1, 2, and 3 would always result in
build 4, but wasn't cached (we previously only cached the build that
existed). Now we also cache the negative case.

Previously, bisecting 4..0 would attempt to find a build for 2. The
closest build is 4, and hence we'd skip an iteration and continue
bisecting 2..0. But the lookup for build 1 also resulted in build 4,
which led to an infinite loop in the old logic.

Bug: 711249
Change-Id: I7608df2ad43e26ecd892b99fd628c4c9040cf60d
Reviewed-on: https://chromium-review.googlesource.com/1163520
Reviewed-by: Sergiy Byelozyorov <sergiyb@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>