Set pending_builds metric for all builders on the master
Without this change, pending_builds metric for the builders that have no pending
builds is not being set and defaults to last known value, which is typically 1.
As a result, V8 alert configs produce false positives as they try to detect when
a builder has pending_builds, but no current_builds.

R=dsansome@chromium.org

Change-Id: I04e787e4f5bcc91dade7abdf3fa1f99cbb5282a6
Reviewed-on: https://chromium-review.googlesource.com/1172126
Reviewed-by: Scott Lee <ddoman@google.com>