zip_build: Apply the path filter recursively when creating a package archive.
Previously we would only apply it to the direct descendants of the
build directory and not the contents of any of its subdirectories. This
meant that, for example, we ended up packaging the 'obj' and
'thinlto-cache' directories from the build directories for toolchains
other than the default toolchain, despite these directories being
filtered out.

We previously had code that handled the 'initial' directory as a
special case where the filter was also being applied to the files
in that directory. It seems like this code is now dead because the
directory is now named 'initialexe' (?), but in any event, since we
now have a generalization of that code, I've removed it.

Bug: 876316
Change-Id: Ieab27788bf3ca7c7bf970434ac491a053eaa5baf
Reviewed-on: https://chromium-review.googlesource.com/1184302
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Peter Collingbourne <pcc@chromium.org>