[bisect] Make debug info step have predictable behavior.
Previously this recipe would run a 'Debug Info' step without any command
which doesn't have well defined behavior (it should probably be an
error, but seems to slip through the cracks). Some of the recipe
simulations mark this as having a retcode of 1, but this should actually
be impossible, leading to simulations which are (I think) out
of line with production behavior.

In addition to this, the recipe would previously modify the
presentation.logs of steps which were already closed which is a no-op
and thus probably not the intended behavior.

Change this to use python.suceeding_step, which has predictable behavior
and has well defined behavior :).

This is necessary for the referenced bug, which is going to make
presentation.logs inaccessible (with an error) after step closure.

R=robertocn@chromium.org

Bug: 833539
Recipe-Nontrivial-Roll: build_limited_scripts_slave
Change-Id: I6810ffcd1e3f95c202eb5322770cfc3bd19a6238
Reviewed-on: https://chromium-review.googlesource.com/1171881
Reviewed-by: Simon Hatch <simonhatch@chromium.org>
Reviewed-by: David Tu <dtu@chromium.org>
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>